*BUG FIX: There was an error when opening an image from the Chart module, invalid cast.
git-svn-id: https://opendentalsoft.com:23793/svn/opendental/head@3257 098d0599-7f2b-0410-95a5-ad30ac85393a

+++ OpenDentBusiness/Data Interface/Documents.cs
@@ -238,7 +238,7 @@ namespace OpenDentBusiness {
 			}
 			ArrayList retVal=new ArrayList();
 			for(int i=0;i<docNums.Count;i++){
-				retVal.Add((string)hList[(int)docNums[i]]);
+				retVal.Add((string)hList[(long)docNums[i]]);
 			}
 			return retVal;
 		}
